refactor(android): move ConnectionTracker and MainProcessConnectionTracker to services/core (#201)
Both types belong to the core layer, not the foreground service layer:
- ConnectionTracker and MainProcessConnectionTracker are the state backend
  for InProcessCallkeepCore, which lives in services/core.
- The foreground service accesses the tracker only indirectly through
  CallkeepCore.instance, it does not own or construct the tracker.
- Moving eliminates the dependency inversion where services/core imported
  from services/services/foreground.

Updated:
- Package declarations in ConnectionTracker.kt and MainProcessConnectionTracker.kt
- Removed now-unnecessary cross-package imports in InProcessCallkeepCore.kt
- Moved MainProcessConnectionTrackerTest.kt to matching services/core test package
- Updated file paths in docs/connection-tracker.md
